Extract ClassMethods constant into autoloadable file
Otherwise, patching it does not work.
